backend/requirements.txt
Django==2.2.4
django-cors-headers==3.1.0
djangorestframework==3.10.2
djangorestframework-jwt==1.11.0
et-xmlfile==1.0.1
jdcal==1.4.1
numpy==1.17.0
openpyxl==2.6.3
pandas==0.25.1
Pillow==6.2.0
pkg-resources==0.0.0
PyJWT==1.7.1
python-dateutil==2.8.0
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
xlrd==1.2.0
